break out HealthMetric and HealthMetricProvider into shared library target/product, draft implementation of HealthMetrics for permissions health (BluetoothManager and NetworkManager)
