+++ Sources/DittoHeartbeat/HeartbeatVM.swift
@@ -9,6 +9,7 @@
 import Combine
 import CryptoKit
 import DittoSwift
+import DittoToolsSharedModels
 import SwiftUI
