+++ Sources/DittoPermissionsHealth/PermissionsHealth.swift
@@ -1,52 +1,12 @@
 //
-//  SwiftUIView.swift
+//  PermissionsHealth.swift
 //
 //
 //  Created by Walker Erekson on 2/26/24.
 //
 
 import SwiftUI
-import CoreBluetooth
 import Combine
-
-@available(iOS 13.0, *)
-class BluetoothManager: NSObject, ObservableObject {
-    private var centralManager: CBCentralManager!
-    private var cancellables = Set<AnyCancellable>()
-
-    @Published var authorizationStatus: CBManagerAuthorization = .notDetermined
-    @Published var managerState: CBManagerState = .unknown
-
-    override init() {
-        super.init()
-        
-        centralManager = CBCentralManager(delegate: self, queue: nil)
-        
-        // Watch for changes in authorization status
-        centralManager.publisher(for: \.authorization)
-            .sink { [weak self] authorization in
-                self?.authorizationStatus = authorization
-            }
-            .store(in: &cancellables)
-        
-        // Watch for changes in manager state
-        centralManager.publisher(for: \.state)
-            .sink { [weak self] state in
-                self?.managerState = state
-            }
-            .store(in: &cancellables)
-    }
-}
-
-@available(iOS 13.0, *)
-extension BluetoothManager: CBCentralManagerDelegate {
-    func centralManagerDidUpdateState(_ central: CBCentralManager) {
-        // This delegate method is called when there's a change in the manager's state.
-        // We don't need to do anything here since we're observing the state using Combine.
-    }
-}
-
-
 
 @available(iOS 14.0, *)
 public struct PermissionsHealth: View {
@@ -67,13 +27,13 @@ public struct PermissionsHealth: View {
                                 .fontWeight(.bold)
                             
                             Spacer()
-                            Image(systemName: checkBluetoothState(state: authorizationStatusDescription) ? "checkmark.circle.fill" : "exclamationmark.triangle.fill")
-                                .foregroundColor(checkBluetoothState(state: authorizationStatusDescription) ? .green : .gray)
+                            Image(systemName: checkBluetoothState(state: bluetoothManager.authorizationStatusDescription) ? "checkmark.circle.fill" : "exclamationmark.triangle.fill")
+                                .foregroundColor(checkBluetoothState(state: bluetoothManager.authorizationStatusDescription) ? .green : .gray)
                                 .font(.title)
                         }
                         Divider()
-                        Text("Permission: \(authorizationStatusDescription)")
-                        if !checkBluetoothState(state: authorizationStatusDescription) {
+                        Text("Permission: \(bluetoothManager.authorizationStatusDescription)")
+                        if !checkBluetoothState(state: bluetoothManager.authorizationStatusDescription) {
                             Divider()
                             Button(action: {
                                 openBluetoothSettings()
@@ -100,15 +60,15 @@ public struct PermissionsHealth: View {
                                 .fontWeight(.bold)
                             
                             Spacer()
-                            Image(systemName: checkBluetoothState(state: managerStateDescription) ? "checkmark.circle.fill" : "exclamationmark.triangle.fill")
-                                .foregroundColor(checkBluetoothState(state: managerStateDescription) ? .green : .gray)
+                            Image(systemName: checkBluetoothState(state: bluetoothManager.managerStateDescription) ? "checkmark.circle.fill" : "exclamationmark.triangle.fill")
+                                .foregroundColor(checkBluetoothState(state: bluetoothManager.managerStateDescription) ? .green : .gray)
                                 .font(.title)
                         }
                         Divider()
-                        Text("Bluetooth: \(managerStateDescription)")
-                        if !checkBluetoothState(state: managerStateDescription) {
+                        Text("Bluetooth: \(bluetoothManager.managerStateDescription)")
+                        if !checkBluetoothState(state: bluetoothManager.managerStateDescription) {
                             Divider()
-                            if(managerStateDescription == "Unauthorized") {
+                            if(bluetoothManager.managerStateDescription == "Unauthorized") {
                                 Text("*See Bluetooth Authorization")
                             } else {
                                 Button(action: {
@@ -163,40 +123,6 @@ public struct PermissionsHealth: View {
                 }
                 .padding()
             }
-        }
-    }
-
-    private var authorizationStatusDescription: String {
-        switch bluetoothManager.authorizationStatus {
-        case .notDetermined:
-            return "Not Determined"
-        case .restricted:
-            return "Restricted"
-        case .denied:
-            return "Denied"
-        case .allowedAlways:
-            return "Allowed Always"
-        @unknown default:
-            return "Unknown"
-        }
-    }
-
-    private var managerStateDescription: String {
-        switch bluetoothManager.managerState {
-        case .unknown:
-            return "Unknown"
-        case .resetting:
-            return "Resetting"
-        case .unsupported:
-            return "Unsupported"
-        case .unauthorized:
-            return "Unauthorized"
-        case .poweredOff:
-            return "Off"
-        case .poweredOn:
-            return "On"
-        @unknown default:
-            return "Unknown"
         }
     }
